requirements.txt
flask
streamlit
pandas
scikit-learn
requests
telethon
google-generativeai
